+++ pkg/ruler/ruler_test.go
@@ -98,19 +98,13 @@ func (r ruleLimits) RulerMaxRulesPerRuleGroup(_ string) int {
 	return r.maxRulesPerRuleGroup
 }
 
-func testSetup(t *testing.T) (*promql.Engine, storage.QueryableFunc, Pusher, log.Logger, RulesLimits) {
-	dir := t.TempDir()
-	tracker := promql.NewActiveQueryTracker(dir, 20, log.NewNopLogger())
-
-	engine := promql.NewEngine(promql.EngineOpts{
-		MaxSamples:         1e6,
-		ActiveQueryTracker: tracker,
-		Timeout:            2 * time.Minute,
-	})
-
+func testSetup() (storage.QueryableFunc, promRules.QueryFunc, Pusher, log.Logger, RulesLimits) {
 	noopQueryable := storage.QueryableFunc(func(ctx context.Context, mint, maxt int64) (storage.Querier, error) {
 		return storage.NoopQuerier(), nil
 	})
+	noopQueryFunc := func(ctx context.Context, q string, t time.Time) (promql.Vector, error) {
+		return nil, nil
+	}
 
 	// Mock the pusher
 	pusher := newPusherMock()
@@ -119,12 +113,14 @@ func testSetup(t *testing.T) (*promql.Engine, storage.QueryableFunc, Pusher, log
 	l := log.NewLogfmtLogger(os.Stdout)
 	l = level.NewFilter(l, level.AllowInfo())
 
-	return engine, noopQueryable, pusher, l, ruleLimits{evalDelay: 0, maxRuleGroups: 20, maxRulesPerRuleGroup: 15}
+	return noopQueryable, noopQueryFunc, pusher, l, ruleLimits{evalDelay: 0, maxRuleGroups: 20, maxRulesPerRuleGroup: 15}
 }
 
 func newManager(t *testing.T, cfg Config) *DefaultMultiTenantManager {
-	engine, noopQueryable, pusher, logger, overrides := testSetup(t)
-	manager, err := NewDefaultMultiTenantManager(cfg, DefaultTenantManagerFactory(cfg, pusher, noopQueryable, noopQueryable, engine, overrides, nil), prometheus.NewRegistry(), logger, nil)
+	noopQueryable, noopQueryFunc, pusher, logger, overrides := testSetup()
+
+	mngFactory := DefaultTenantManagerFactory(cfg, pusher, noopQueryable, noopQueryFunc, overrides, nil)
+	manager, err := NewDefaultMultiTenantManager(cfg, mngFactory, prometheus.NewRegistry(), logger, nil)
 	require.NoError(t, err)
 
 	return manager
@@ -169,10 +165,10 @@ func newMockClientsPool(cfg Config, logger log.Logger, reg prometheus.Registerer
 }
 
 func buildRuler(t *testing.T, cfg Config, storage rulestore.RuleStore, rulerAddrMap map[string]*Ruler) *Ruler {
-	engine, noopQueryable, pusher, logger, overrides := testSetup(t)
+	noopQueryable, noopQueryFunc, pusher, logger, overrides := testSetup()
 
 	reg := prometheus.NewRegistry()
-	managerFactory := DefaultTenantManagerFactory(cfg, pusher, noopQueryable, noopQueryable, engine, overrides, reg)
+	managerFactory := DefaultTenantManagerFactory(cfg, pusher, noopQueryable, noopQueryFunc, overrides, reg)
 	manager, err := NewDefaultMultiTenantManager(cfg, managerFactory, reg, log.NewNopLogger(), nil)
 	require.NoError(t, err)