+++ pkg/ruler/tenant_federation.go
@@ -37,10 +37,18 @@ func FederatedGroupContextFunc(ctx context.Context, g *rules.Group) context.Cont
 	return context.WithValue(ctx, federatedGroupSourceTenants, g.SourceTenants())
 }
 
+// ExtractTenantIDs gets the rule group org ID from the context.
+func ExtractTenantIDs(ctx context.Context) (string, error) {
+	if sourceTenants, _ := ctx.Value(federatedGroupSourceTenants).([]string); len(sourceTenants) > 0 {
+		return tenant.JoinTenantIDs(tenant.NormalizeTenantIDs(sourceTenants)), nil
+	}
+	return tenant.TenantID(ctx)
+}
+
 func TenantFederationQueryFunc(regularQueryable, federatedQueryable rules.QueryFunc) rules.QueryFunc {
 	return func(ctx context.Context, q string, t time.Time) (promql.Vector, error) {
 		if sourceTenants, _ := ctx.Value(federatedGroupSourceTenants).([]string); len(sourceTenants) > 0 {
-			ctx = user.InjectOrgID(ctx, tenant.JoinTenantIDs(sourceTenants))
+			ctx = user.InjectOrgID(ctx, tenant.JoinTenantIDs(tenant.NormalizeTenantIDs(sourceTenants)))
 			return federatedQueryable(ctx, q, t)
 		}
 		return regularQueryable(ctx, q, t)